resizable widgets, group on/off ok

+++ RFGraph_Model.py
@@ -557,7 +557,7 @@ class RFGraph_Model(QtGui.QMainWindow):
 
         self.hide()
         equacolOtmp = np.array(equacolOtmp, dtype=object)
-        equacolOtmp = equacolOtmp.reshape(len(equacolOtmp)/8,8)
+        equacolOtmp = equacolOtmp.reshape(int(len(equacolOtmp)/8),8)
 
 
         return equacolOtmp
